Decouple the deployment manager from verticles.
Motivation:

The deployment manager handle verticle concerns (such as creating contexts, worker pool, etc...) and impacts the testability and readability of deploy/undeploy (lifecycle) operations.

Changes:

Abstract the verticle details in a Deployable interface which is presented to the deployment manager by the verticle manager which already handle partially the handling of the verticle.

Results:

The deployment manager is not anymore coupled to verticles and can be tested / changed independantly.
