+++ vertx-core/src/main/java/io/vertx/core/impl/ContextImpl.java
@@ -14,6 +14,7 @@ package io.vertx.core.impl;
 import io.netty.channel.EventLoop;
 import io.vertx.core.*;
 import io.vertx.core.Future;
+import io.vertx.core.impl.deployment.Deployment;
 import io.vertx.core.internal.EventExecutor;
 import io.vertx.core.internal.logging.Logger;
 import io.vertx.core.internal.logging.LoggerFactory;
