--- vertx-core/src/main/java/io/vertx/core/impl/DeploymentManager.java
@@ -1,470 +1,0 @@
-/*
- * Copyright (c) 2011-2019 Contributors to the Eclipse Foundation
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
- * which is available at https://www.apache.org/licenses/LICENSE-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
- */
-
-package io.vertx.core.impl;
-
-import io.netty.channel.EventLoop;
-import io.vertx.core.*;
-import io.vertx.core.internal.CloseFuture;
-import io.vertx.core.internal.ContextInternal;
-import io.vertx.core.json.JsonObject;
-import io.vertx.core.internal.logging.Logger;
-import io.vertx.core.internal.logging.LoggerFactory;
-
-import java.util.*;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Function;
-
-/**
- * @author <a href="http://tfox.org">Tim Fox</a>
- */
-public class DeploymentManager {
-
-  private static final Logger log = LoggerFactory.getLogger(DeploymentManager.class);
-
-  private final VertxImpl vertx;
-  private final Map<String, Deployment> deploying = new HashMap<>();
-  private final Map<String, Deployment> deployments = new ConcurrentHashMap<>();
-
-  public DeploymentManager(VertxImpl vertx) {
-    this.vertx = vertx;
-  }
-
-  private String generateDeploymentID() {
-    return UUID.randomUUID().toString();
-  }
-
-  public Future<String> deployVerticle(Callable<Verticle> verticleSupplier, DeploymentOptions options) {
-    if (options.getInstances() < 1) {
-      throw new IllegalArgumentException("Can't specify < 1 instances to deploy");
-    }
-    options.checkIsolationNotDefined();
-    ContextInternal currentContext = vertx.getOrCreateContext();
-    ClassLoader cl = options.getClassLoader();
-    if (cl == null) {
-      cl = Thread.currentThread().getContextClassLoader();
-      if (cl == null) {
-        cl = getClass().getClassLoader();
-      }
-    }
-    return doDeploy(options, v -> "java:" + v.getClass().getName(), currentContext, currentContext, cl, verticleSupplier)
-      .map(Deployment::deploymentID);
-  }
-
-  public Future<Void> undeployVerticle(String deploymentID) {
-    Deployment deployment = deployments.get(deploymentID);
-    ContextInternal currentContext = vertx.getOrCreateContext();
-    if (deployment == null) {
-      return currentContext.failedFuture(new IllegalStateException("Unknown deployment"));
-    } else {
-      return deployment.doUndeploy(vertx.getOrCreateContext());
-    }
-  }
-
-  public Set<String> deployments() {
-    return Collections.unmodifiableSet(deployments.keySet());
-  }
-
-  public Deployment getDeployment(String deploymentID) {
-    return deployments.get(deploymentID);
-  }
-
-  public Future<Void> undeployAll() {
-    // TODO timeout if it takes too long - e.g. async stop verticle fails to call future
-
-    // We only deploy the top level verticles as the children will be undeployed when the parent is
-    while (true) {
-      DeploymentImpl deployment;
-      synchronized (deploying) {
-        if (deploying.isEmpty()) {
-          break;
-        }
-        Iterator<Map.Entry<String, Deployment>> it = deploying.entrySet().iterator();
-        Map.Entry<String, Deployment> entry = it.next();
-        it.remove();
-        deployment = (DeploymentImpl) entry.getValue();
-      }
-      deployment.verticles.forEach(holder -> {
-        Promise<Void> startPromise = holder.startPromise;
-        if (startPromise != null) {
-          startPromise.tryFail(new VertxException("Verticle undeployed", true));
-        }
-      });
-    }
-    Set<String> deploymentIDs = new HashSet<>();
-    for (Map.Entry<String, Deployment> entry: deployments.entrySet()) {
-      if (!entry.getValue().isChild()) {
-        deploymentIDs.add(entry.getKey());
-      }
-    }
-    List<Future<?>> completionList = new ArrayList<>();
-    if (!deploymentIDs.isEmpty()) {
-      for (String deploymentID : deploymentIDs) {
-        Promise<Void> promise = Promise.promise();
-        completionList.add(promise.future());
-        undeployVerticle(deploymentID).onComplete(ar -> {
-          if (ar.failed()) {
-            // Log but carry on regardless
-            log.error("Undeploy failed", ar.cause());
-          }
-          promise.handle(ar);
-        });
-      }
-      Promise<Void> promise = vertx.getOrCreateContext().promise();
-      Future.join(completionList).<Void>mapEmpty().onComplete(promise);
-      return promise.future();
-    } else {
-      return vertx.getOrCreateContext().succeededFuture();
-    }
-  }
-
-  Future<Deployment> doDeploy(DeploymentOptions options,
-                                  Function<Verticle, String> identifierProvider,
-                                  ContextInternal parentContext,
-                                  ContextInternal callingContext,
-                                  ClassLoader tccl, Callable<Verticle> verticleSupplier) {
-    int nbInstances = options.getInstances();
-    Set<Verticle> verticles = Collections.newSetFromMap(new IdentityHashMap<>());
-    for (int i = 0; i < nbInstances; i++) {
-      Verticle verticle;
-      try {
-        verticle = verticleSupplier.call();
-      } catch (Exception e) {
-        return Future.failedFuture(e);
-      }
-      if (verticle == null) {
-        return Future.failedFuture("Supplied verticle is null");
-      }
-      verticles.add(verticle);
-    }
-    if (verticles.size() != nbInstances) {
-      return Future.failedFuture("Same verticle supplied more than once");
-    }
-    Verticle[] verticlesArray = verticles.toArray(new Verticle[0]);
-    return doDeploy(identifierProvider.apply(verticlesArray[0]), options, parentContext, callingContext, tccl, verticlesArray);
-  }
-
-  private Future<Deployment> doDeploy(String identifier,
-                                      DeploymentOptions options,
-                                      ContextInternal parentContext,
-                                      ContextInternal callingContext,
-                                      ClassLoader tccl,
-                                      Verticle... verticles) {
-    Promise<Deployment> promise = callingContext.promise();
-    Deployment parent = parentContext.getDeployment();
-    String deploymentID = generateDeploymentID();
-
-    AtomicInteger deployCount = new AtomicInteger();
-    AtomicBoolean failureReported = new AtomicBoolean();
-    WorkerPool workerPool = null;
-    ThreadingModel mode = options.getThreadingModel();
-    if (mode == null) {
-      mode = ThreadingModel.EVENT_LOOP;
-    }
-    if (mode != ThreadingModel.VIRTUAL_THREAD) {
-      if (options.getWorkerPoolName() != null) {
-        workerPool = vertx.createSharedWorkerPool(options.getWorkerPoolName(), options.getWorkerPoolSize(), options.getMaxWorkerExecuteTime(), options.getMaxWorkerExecuteTimeUnit());
-      }
-    } else {
-      if (!vertx.isVirtualThreadAvailable()) {
-        return callingContext.failedFuture("This Java runtime does not support virtual threads");
-      }
-    }
-    EventLoop workerLoop = null;
-    DeploymentImpl deployment = new DeploymentImpl(parent, workerPool, deploymentID, identifier, options);
-    synchronized (deploying) {
-      deploying.put(deploymentID, deployment);
-    }
-    for (Verticle verticle: verticles) {
-      CloseFuture closeFuture = new CloseFuture(log);
-      ContextImpl context;
-      switch (mode) {
-        default:
-          context = vertx.createEventLoopContext(deployment, closeFuture, workerPool, tccl);
-          break;
-        case WORKER:
-          if (workerLoop == null) {
-            context = vertx.createWorkerContext(deployment, closeFuture, workerPool, tccl);
-            workerLoop = context.nettyEventLoop();
-          } else {
-            context = vertx.createWorkerContext(deployment, closeFuture, workerLoop, workerPool, tccl);
-          }
-          break;
-        case VIRTUAL_THREAD:
-          if (workerLoop == null) {
-            context = vertx.createVirtualThreadContext(deployment, closeFuture, tccl);
-            workerLoop = context.nettyEventLoop();
-          } else {
-            context = vertx.createVirtualThreadContext(deployment, closeFuture, workerLoop, tccl);
-          }
-          break;
-      }
-      VerticleHolder holder = new VerticleHolder(verticle, context, closeFuture);
-      Promise<Void> startPromise = context.promise();
-      holder.startPromise = startPromise;
-      deployment.addVerticle(holder);
-      context.runOnContext(v -> {
-        try {
-          verticle.init(vertx, context);
-          Future<Void> startFuture = startPromise.future();
-          verticle.start(startPromise);
-          startFuture.onComplete(ar -> {
-            holder.startPromise = null;
-            if (ar.succeeded()) {
-              if (parent != null) {
-                if (parent.addChild(deployment)) {
-                  deployment.child = true;
-                } else {
-                  // Orphan
-                  deployment.doUndeploy(vertx.getOrCreateContext()).onComplete(ar2 -> promise.fail("Verticle deployment failed.Could not be added as child of parent verticle"));
-                  return;
-                }
-              }
-              deployments.put(deploymentID, deployment);
-              if (deployCount.incrementAndGet() == verticles.length) {
-                synchronized (deploying) {
-                  deploying.remove(deploymentID);
-                }
-                promise.complete(deployment);
-              }
-            } else if (failureReported.compareAndSet(false, true)) {
-              deployment.rollback(callingContext, promise, context, holder, ar.cause());
-            }
-          });
-        } catch (Throwable t) {
-          if (failureReported.compareAndSet(false, true))
-            deployment.rollback(callingContext, promise, context, holder, t);
-        }
-      });
-    }
-
-    return promise.future();
-  }
-
-  static class VerticleHolder {
-
-    final Verticle verticle;
-    final ContextImpl context;
-    final CloseFuture closeFuture;
-    Promise<Void> startPromise;
-
-    VerticleHolder(Verticle verticle, ContextImpl context, CloseFuture closeFuture) {
-      this.verticle = verticle;
-      this.context = context;
-      this.closeFuture = closeFuture;
-    }
-
-    Future<Void> close() {
-      return closeFuture.close();
-    }
-  }
-
-  private class DeploymentImpl implements Deployment {
-
-    private static final int ST_DEPLOYED = 0, ST_UNDEPLOYING = 1, ST_UNDEPLOYED = 2;
-
-    private final Deployment parent;
-    private final String deploymentID;
-    private final JsonObject conf;
-    private final String verticleIdentifier;
-    private final List<VerticleHolder> verticles = new CopyOnWriteArrayList<>();
-    private final Set<Deployment> children = ConcurrentHashMap.newKeySet();
-    private final WorkerPool workerPool;
-    private final DeploymentOptions options;
-    private Handler<Void> undeployHandler;
-    private int status = ST_DEPLOYED;
-    private volatile boolean child;
-
-    private DeploymentImpl(Deployment parent, WorkerPool workerPool, String deploymentID, String verticleIdentifier, DeploymentOptions options) {
-      this.parent = parent;
-      this.deploymentID = deploymentID;
-      this.conf = options.getConfig() != null ? options.getConfig().copy() : new JsonObject();
-      this.verticleIdentifier = verticleIdentifier;
-      this.options = options;
-      this.workerPool = workerPool;
-    }
-
-    public void addVerticle(VerticleHolder holder) {
-      verticles.add(holder);
-    }
-
-    private synchronized void rollback(ContextInternal callingContext, Promise<Deployment> completionPromise, ContextImpl context, VerticleHolder closeFuture, Throwable cause) {
-      if (status == ST_DEPLOYED) {
-        status = ST_UNDEPLOYING;
-        doUndeployChildren(callingContext).onComplete(childrenResult -> {
-          if (workerPool != null) {
-            workerPool.close();
-          }
-          Handler<Void> handler;
-          synchronized (DeploymentImpl.this) {
-            status = ST_UNDEPLOYED;
-            handler = undeployHandler;
-            undeployHandler = null;
-          }
-          if (handler != null) {
-            try {
-              handler.handle(null);
-            } catch (Exception e) {
-              context.reportException(e);
-            }
-          }
-          if (childrenResult.failed()) {
-            completionPromise.fail(cause);
-          } else {
-            closeFuture.close().transform(ar -> Future.<Deployment>failedFuture(cause)).onComplete(completionPromise);
-          }
-        });
-      }
-    }
-
-    private synchronized Future<Void> doUndeployChildren(ContextInternal undeployingContext) {
-      if (!children.isEmpty()) {
-        List<Future<?>> childFuts = new ArrayList<>();
-        for (Deployment childDeployment: new HashSet<>(children)) {
-          Promise<Void> p = Promise.promise();
-          childFuts.add(p.future());
-          childDeployment.doUndeploy(undeployingContext).onComplete(ar -> {
-            children.remove(childDeployment);
-            p.handle(ar);
-          });
-        }
-        return Future.all(childFuts).mapEmpty();
-      } else {
-        return Future.succeededFuture();
-      }
-    }
-
-    public synchronized Future<Void> doUndeploy(ContextInternal undeployingContext) {
-      if (status == ST_UNDEPLOYED) {
-        return Future.failedFuture(new IllegalStateException("Already undeployed"));
-      }
-      if (!children.isEmpty()) {
-        status = ST_UNDEPLOYING;
-        return doUndeployChildren(undeployingContext).compose(v -> doUndeploy(undeployingContext));
-      } else {
-        status = ST_UNDEPLOYED;
-        List<Future<?>> undeployFutures = new ArrayList<>();
-        if (parent != null) {
-          parent.removeChild(this);
-        }
-        for (VerticleHolder verticleHolder: verticles) {
-          ContextImpl context = verticleHolder.context;
-          Promise<Void> p = Promise.promise();
-          undeployFutures.add(p.future());
-          context.runOnContext(v -> {
-            Promise<Void> stopPromise = undeployingContext.promise();
-            Future<Void> stopFuture = stopPromise.future();
-            stopFuture
-              .eventually(() -> {
-                deployments.remove(deploymentID);
-                return verticleHolder
-                  .close()
-                  .onFailure(err -> log.error("Failed to run close hook", err));
-              }).onComplete(p);
-            try {
-              verticleHolder.verticle.stop(stopPromise);
-            } catch (Throwable t) {
-              if (!stopPromise.tryFail(t)) {
-                undeployingContext.reportException(t);
-              }
-            }
-          });
-        }
-        Promise<Void> resolvingPromise = undeployingContext.promise();
-        Future.all(undeployFutures).<Void>mapEmpty().onComplete(resolvingPromise);
-        Future<Void> fut = resolvingPromise.future();
-        if (workerPool != null) {
-          fut = fut.andThen(ar -> workerPool.close());
-        }
-        Handler<Void> handler = undeployHandler;
-        if (handler != null) {
-          undeployHandler = null;
-          return fut.andThen(ar -> handler.handle(null));
-        }
-        return fut;
-      }
-    }
-
-    @Override
-    public String verticleIdentifier() {
-      return verticleIdentifier;
-    }
-
-    @Override
-    public DeploymentOptions deploymentOptions() {
-      return options;
-    }
-
-    @Override
-    public JsonObject config() {
-      return conf;
-    }
-
-    @Override
-    public synchronized boolean addChild(Deployment deployment) {
-      if (status == ST_DEPLOYED) {
-        children.add(deployment);
-        return true;
-      } else {
-        return false;
-      }
-    }
-
-    @Override
-    public void removeChild(Deployment deployment) {
-      children.remove(deployment);
-    }
-
-    @Override
-    public Set<Context> getContexts() {
-      Set<Context> contexts = new HashSet<>();
-      for (VerticleHolder holder: verticles) {
-        contexts.add(holder.context);
-      }
-      return contexts;
-    }
-
-    @Override
-    public Set<Verticle> getVerticles() {
-      Set<Verticle> verts = new HashSet<>();
-      for (VerticleHolder holder: verticles) {
-        verts.add(holder.verticle);
-      }
-      return verts;
-    }
-
-    @Override
-    public void undeployHandler(Handler<Void> handler) {
-      synchronized (this) {
-        if (status != ST_UNDEPLOYED) {
-          undeployHandler = handler;
-          return;
-        }
-      }
-      handler.handle(null);
-    }
-
-    @Override
-    public boolean isChild() {
-      return child;
-    }
-
-    @Override
-    public String deploymentID() {
-      return deploymentID;
-    }
-  }
-
-}
