+++ vertx-core/src/main/java/io/vertx/core/impl/DuplicatedContext.java
@@ -15,6 +15,7 @@ import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.ThreadingModel;
+import io.vertx.core.impl.deployment.Deployment;
 import io.vertx.core.internal.CloseFuture;
 import io.vertx.core.internal.ContextInternal;
 import io.vertx.core.internal.EventExecutor;
