+++ vertx-core/src/main/java/io/vertx/core/impl/deployment/Deployable.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2011-2024 Contributors to the Eclipse Foundation
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+ * which is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+ */
+package io.vertx.core.impl.deployment;
+
+import io.vertx.core.Future;
+import io.vertx.core.Promise;
+
+/**
+ * Something we can deploy.
+ */
+public interface Deployable {
+
+  String identifier();
+
+  void deploy(Deployment deployment, Promise<Void> completion);
+
+  Future<?> undeploy();
+
+  Future<?> cleanup();
+
+}
