+++ vertx-core/src/main/java/io/vertx/core/internal/ContextInternal.java
@@ -15,6 +15,7 @@ import io.netty.channel.EventLoop;
 import io.vertx.core.*;
 import io.vertx.core.Future;
 import io.vertx.core.impl.*;
+import io.vertx.core.impl.deployment.Deployment;
 import io.vertx.core.impl.future.FailedFuture;
 import io.vertx.core.impl.future.PromiseImpl;
 import io.vertx.core.impl.future.SucceededFuture;
