+++ vertx-core/src/main/java/io/vertx/core/internal/VertxWrapper.java
@@ -23,6 +23,7 @@ import io.vertx.core.eventbus.EventBus;
 import io.vertx.core.file.FileSystem;
 import io.vertx.core.http.*;
 import io.vertx.core.impl.*;
+import io.vertx.core.impl.deployment.Deployment;
 import io.vertx.core.internal.threadchecker.BlockedThreadChecker;
 import io.vertx.core.net.NetClient;
 import io.vertx.core.net.NetClientOptions;
